backend/requirements.txt
fastapi==0.123.9
uvicorn[standard]==0.23.2
pydantic==2.12.2
pydantic-settings==2.4.0
sqlalchemy==2.0.35
aiosqlite==0.20.0
greenlet==3.1.1
alembic==1.17.2
redis==6.4.0
python-jose[cryptography]==3.5.0
python-multipart==0.0.8
solana==0.36.10
solders==0.27.1
httpx==0.28.1
python-dotenv==1.2.1
slowapi==0.1.9
